backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "edge-tts>=7.2.7",
    "email-validator>=2.3.0",
    "fastapi>=0.133.0",
    "langchain-core>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "python-ulid>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "upstash-redis>=1.6.0",
    "uvicorn>=0.41.0",
]
